requirements.txt
Django>=5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
pillow>=10.0.0  
django-ninja>=1.0.0
